feat: Download data from the Bulk Data Service
https://github.com/IATI/IATI-Dashboard/issues/768

This replaces the old data folder from IATI-Registry-Refresher, and the
metadata that had been downloaded from ckan.

The `wrong_roots` check no longer checks the metadata, as we don't
record whether a file is meant to be activities or organisations (we
just detect it later based on the root). With this, nothing in
IATI-Stats uses the metadata, but it is plaecd in the output directory
for IATI-Dashboard to use.

Whilst I'm editing git.sh, I swtiched it to use set-eux:
https://github.com/IATI/IATI-Dashboard/issues/757